// src/app/auth/models/interfaces/login.interface.ts
// Generated by https://quicktype.io

export interface IUserLoginRequest {
  username: string;
  password: string;
}

// Generated by https://quicktype.io

export interface ILoginResponse {
  id: number;
  username: string;
  correoElectronico: string;
  accesToken: string;
  refreshToken: string;
}

/* export interface IRefreshTokenResponse {
  username: string;
  refreshToken: string;
} */

// Generated by https://quicktype.io

export interface IRefreshTokenResponse {
  accessToken:  string;
  refreshToken: string;
}


export interface IRefreshTokenRequest {
  username: string;
  refreshToken: string;
}